Move registerUser to functions
